Move CAC to a separate config folder
