+++ .github/workflows/deploy-config.yml
@@ -5,12 +5,7 @@ on:
     branches:
       - main
     paths:
-      - "docker-compose.prod.yml"
-      - "Caddyfile.prod"
-      - "litestream.yml"
-      - "Dockerfile.caddy"
-      - "scripts/bindersnap"
-      - "scripts/bootstrap-gitea-service-account.ts"
+      - "config/**"
   workflow_dispatch:
     inputs:
       publish_from_repo:
@@ -104,7 +99,7 @@ jobs:
         run: |
           set -euo pipefail
 
-          aws s3 sync ./ "s3://${CONFIG_BUCKET}/" \
+          aws s3 sync ./config/ "s3://${CONFIG_BUCKET}/" \
             --exclude "*" \
             --include "docker-compose.prod.yml" \
             --include "Caddyfile.prod" \
