+++ config/scripts/bootstrap-gitea-service-account.ts
