+++ scripts/bootstrap-gitea-service-account.test.ts
@@ -9,7 +9,7 @@ import {
   resolveBootstrapConfig,
   resolveServiceTokenScopes,
   resolveSsmParameterName,
-} from "./bootstrap-gitea-service-account";
+} from "../config/scripts/bootstrap-gitea-service-account";
 
 describe("bootstrap-gitea-service-account", () => {
   test("defaults to the minimum admin scope required by the API", () => {
